restruction application code and add to index.md
